requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.12
albumentations==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==21.4.0
audioread==3.0.1
Babel==2.10.1
backports.entry-points-selectable==1.1.1
backports.functools-lru-cache==1.6.4
bcolz-zipline==1.2.9
bcrypt==3.2.2
beautifulsoup4==4.12.3
bitstring==3.1.9
blinker==1.7.0
blist==1.3.6
CacheControl==0.12.11
cachetools==5.3.2
cachy==0.3.0
certifi==2023.11.17
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.3.2
cleo==0.8.1
click==8.1.7
clikit==0.6.2
cmake==3.29.3
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.0
crashtest==0.3.1
cryptography==37.0.1
cycler==0.12.1
Cython==0.29.28
datasets==3.0.0
decorator==5.1.1
deepface==0.0.90
diffusers==0.23.0
diffwave==0.1.8
dill==0.3.8
distlib==0.3.4
dnspython==2.6.1
docopt==0.6.2
docutils==0.17.1
easydict==1.13
ecdsa==0.17.0
editables==0.3
email_validator==2.1.1
exceptiongroup==1.2.1
facenet-pytorch==2.5.3
fastapi==0.111.0
fastapi-cli==0.0.3
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
Flask==3.0.1
flatbuffers==23.5.26
flit==3.7.1
flit_core==3.7.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.2
gast==0.5.4
gdown==5.0.0
glob2==0.7
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
gradio==4.8.0
gradio_client==0.7.1
graphviz==0.8.4
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.4.3
imagesize==1.3.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==1.1.1
insightface==0.7.3
intervaltree==3.1.0
intreehooks==1.0
ipaddress==1.0.23
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.4.0
keras==2.15.0
keyring==23.5.0
keyrings.alt==4.1.0
kiwisolver==1.4.5
lazy_loader==0.4
liac-arff==2.5.0
libclang==16.0.6
librosa==0.10.2.post1
lightning-utilities==0.11.7
lit==18.1.4
llvmlite==0.43.0
lockfile==0.12.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.9.0
mdurl==0.1.2
menpo==0.11.1
ml-dtypes==0.2.0
mock==4.0.3
more-itertools==8.12.0
mpmath==1.3.0
msgpack==1.0.3
mtcnn==0.1.1
multidict==6.1.0
multiprocess==0.70.16
mxnet==1.9.1
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.0
onnxruntime-gpu==1.17.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.3
packaging==23.2
pandas==2.2.0
paramiko==2.10.4
pastel==0.2.1
pathlib2==2.3.7.post1
pathspec==0.9.0
pbr==5.8.1
peft==0.10.0
pexpect==4.8.0
pillow==10.2.0
pip==22.0.4
pkginfo==1.8.2
platformdirs==2.4.1
pluggy==1.0.0
poetry==1.1.13
poetry-core==1.0.8
prettytable==3.10.0
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
py-expression-eval==0.3.14
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycrypto==2.6.1
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pylev==1.4.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyrsistent==0.18.1
PySocks==1.7.1
pyspng==0.1.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytoml==0.1.21
pytorch-ignite==0.5.1
pytorch-lightning==1.8.6
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
regex==2022.4.24
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retina-face==0.0.13
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scandir==1.10.0
scikit-image==0.23.2
scikit-learn==1.2.2
scipy==1.13.0
SecretStorage==3.3.2
semantic-version==2.9.0
setuptools==62.1.0
setuptools-rust==1.3.0
setuptools-scm==6.4.2
shellingham==1.4.0
simplegeneric==0.8.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
Sphinx==4.5.0
sphinx-bootstrap-theme==0.8.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
starlette==0.37.2
sympy==1.12
tabulate==0.8.9
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.4.0
threadpoolctl==3.1.0
tifffile==2024.5.10
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchmetrics==1.4.1
torchvision==0.15.2
tqdm==4.66.4
transformers==4.34.1
triton==2.3.1
typer==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.2.0
urllib3==2.1.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.14.1
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.37.1
wrapt==1.14.1
xlrd==2.0.1
xxhash==3.5.0
yarl==1.11.1
zipfile36==0.1.3
zipp==3.17.0
